// src/api/generated/model/sessionCollectionItemGameMode.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */

/**
 * Session game mode
 */
export type SessionCollectionItemGameMode =
  (typeof SessionCollectionItemGameMode)[keyof typeof SessionCollectionItemGameMode];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SessionCollectionItemGameMode = {
  ONE_V_ONE: 'ONE_V_ONE',
  TWO_V_TWO: 'TWO_V_TWO',
  THREE_V_THREE: 'THREE_V_THREE',
  FOUR_V_FOUR: 'FOUR_V_FOUR',
  FIVE_V_FIVE: 'FIVE_V_FIVE',
  SIX_V_SIX: 'SIX_V_SIX',
  SEVEN_V_SEVEN: 'SEVEN_V_SEVEN',
  EIGHT_V_EIGHT: 'EIGHT_V_EIGHT',
  TEN_V_TEN: 'TEN_V_TEN',
  ELEVEN_V_ELEVEN: 'ELEVEN_V_ELEVEN',
} as const;