// src/api/generated/model/sessionCollectionItemSport.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API v1
 * API for the Ludora app
 * OpenAPI spec version: 1.0
 */

/**
 * Sport played during the session
 */
export type SessionCollectionItemSport = (typeof SessionCollectionItemSport)[keyof typeof SessionCollectionItemSport];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SessionCollectionItemSport = {
  BASKETBALL: 'BASKETBALL',
  FOOTBALL: 'FOOTBALL',
  TENNIS: 'TENNIS',
  VOLLEYBALL: 'VOLLEYBALL',
  PADDEL: 'PADDEL',
  BADMINTON: 'BADMINTON',
  'PING-PONG': 'PING-PONG',
} as const;